more user functionslity

+++ app/router/post.py
@@ -13,11 +13,17 @@ app = APIRouter(
 
 # post
 @app.post("/", response_model = posts.PostOut, status_code = 201)
-async def post(post_data: posts.CreatePost, db: Session = Depends(get_session)):
+async def post(post_data: posts.CreatePost, db: Session = Depends(get_session), user: users.User= Depends(Token_Data.get_current_user)):
 
     # if not member logic 
     
+    temp_data = post_data.dict()
+    temp_data={**temp_data, "owner_email":user.email}
+   
+    post_data = posts.FinalCreation(**temp_data)
+   
     post_data = posts.Posts.from_orm(post_data)
+    
     
     db.add(post_data)
     await db.commit()
@@ -26,7 +32,7 @@ async def post(post_data: posts.CreatePost, db: Session = Depends(get_session)):
 
 # get all post
 @app.get("/", response_model = List[posts.PostOut])
-async def get_all(db: Session = Depends(get_session)):
+async def get_all(db: Session = Depends(get_session), user: users.User = Depends(Token_Data.get_current_user)):
     
     data = await db.exec(select(posts.Posts))
 
@@ -36,9 +42,10 @@ async def get_all(db: Session = Depends(get_session)):
 
     return data
 
+
 # get post by id
 @app.get("/{id}", response_model = posts.PostOut)
-async def get_id(id: int, db: Session = Depends(get_session)):
+async def get_id(id: int, db: Session = Depends(get_session), user: users.User = Depends(Token_Data.get_current_user)):
     
     data = await db.exec(select(posts.Posts).where(posts.Posts.id == id))
     data: posts.Posts | None = data.first()
@@ -48,21 +55,26 @@ async def get_id(id: int, db: Session = Depends(get_session)):
 
 #delete post
 @app.delete("/{id}", status_code = 204)
-async def delete_post(id: int, db: Session = Depends(get_session)):
+async def delete_post(id: int, db: Session = Depends(get_session), user: users.User = Depends(Token_Data.get_current_user)):
+    
     data: posts.Posts | None = await db.get(posts.Posts, id)
     if not data:
         raise HTTPException(status_code = 404, detail = f"No data with an id {id}")
+    if user.email != data.owner_email:
+        raise HTTPException(status_code=403, detail ="You are not the creator of this post")
     await db.delete(data)
     await db.commit()
     return Response(status_code = 204)
 
 #update post
-@app.put("/{id}")
-async def update(id:int, post_data: posts.UpdatePost, db: Session= Depends(get_session)):
-    
+@app.put("/{id}", response_model = posts.PostOut)
+async def update(id:int, post_data: posts.UpdatePost, db: Session= Depends(get_session), user: users.User = Depends(Token_Data.get_current_user)):
+
     data: posts.Posts | None = await db.get(posts.Posts, id)
     if not data:
         raise HTTPException(status_code = 404, detail = f"No data with an id {id}")
+    if user.email != data.owner_email:
+        raise HTTPException(status_code=403, detail ="You are not the creator of this post")
     post = post_data.dict(exclude_unset = True)
     
     for key, value in post.items():
